# Cache `get_obj_from_pointers` call for function.